Dutch College Dataset from the amen Package

This is the dutchcollege dataset from the amen package in wide format and was used in van de Bunt, van Duijn and Snijders (1999).

data(data.dc)

Format

A data frame with 992 observations on the following 10 variables. The variables T1, ..., T7 contrain observations from waves 1 to 7.

'data.frame': 992 obs. of 10 variables:
$ Actor : int 1 2 1 3 1 4 1 5 1 6 ...
$ Partner: int 2 1 3 1 4 1 5 1 6 1 ...
$ dyad : int 1 1 2 2 3 3 4 4 5 5 ...
$ T1 : num 0 0 0 0 0 0 0 0 0 0 ...
$ T2 : num 0 0 0 0 0 0 0 0 0 0 ...
$ T3 : num 1 0 0 0 0 0 0 0 0 0 ...
$ T4 : num 1 0 0 0 0 0 0 0 0 0 ...
$ T5 : num 1 0 0 3 0 0 0 0 0 0 ...
$ T6 : num 1 0 0 0 0 0 0 0 0 0 ...
$ T7 : num 1 0 0 0 0 0 0 0 0 0 ...

Source

data(dutchcollege, package="amen")

References

van de Bunt, G. G., Van Duijn, M. A., & Snijders, T. A. (1999). Friendship networks through time: An actor-oriented dynamic statistical network model. Computational & Mathematical Organization Theory, 5(2), 167-192. doi: 10.1023/A:1009683123448